Bluetooth: Reduce bt_keys size
Makes bt_keys struct more compact saving 4 bytes.


Former-commit-id: e6bc2ebc987e9bcdd49fb5c74032920ccea3e520

+++ net/bluetooth/keys.h
@@ -17,7 +17,7 @@ DECL|ediv|member|uint16_t ediv;
 DECL|enc_size|member|uint8_t enc_size;
 DECL|flags|member|atomic_t flags;
 DECL|irk|member|struct bt_irk irk;
-DECL|keys|member|int keys;
+DECL|keys|member|uint16_t keys;
 DECL|local_csrk|member|struct bt_csrk local_csrk;
 DECL|ltk|member|struct bt_ltk ltk;
 DECL|rand|member|uint64_t rand;
